refactor: remove DTO name in events
